refactor(starknet_os): move patricia related hints to the patricia hints crate (#6050)
