+++ crates/starknet_os/src/hints/enum_definition.rs
@@ -82,13 +82,6 @@ use crate::hints::hint_implementation::execution::implementation::{
     end_tx,
     enter_call,
     enter_scope_deprecated_syscall_handler,
-    enter_scope_descend_edge,
-    enter_scope_left_child,
-    enter_scope_new_node,
-    enter_scope_next_node_bit_0,
-    enter_scope_next_node_bit_1,
-    enter_scope_node,
-    enter_scope_right_child,
     enter_scope_syscall_handler,
     enter_syscall_scopes,
     exit_call,
@@ -156,8 +149,18 @@ use crate::hints::hint_implementation::output::{
 use crate::hints::hint_implementation::patricia::implementation::{
     assert_case_is_right,
     build_descent_map,
+    decode_node,
+    enter_scope_descend_edge,
+    enter_scope_left_child,
+    enter_scope_new_node,
+    enter_scope_next_node_bit_0,
+    enter_scope_next_node_bit_1,
+    enter_scope_node,
+    enter_scope_right_child,
     height_is_zero_or_len_node_preimage_is_two,
     is_case_right,
+    load_bottom,
+    load_edge,
     prepare_preimage_validation_non_deterministic_hashes,
     set_ap_to_descend,
     set_bit,
@@ -173,11 +176,8 @@ use crate::hints::hint_implementation::resources::{
 use crate::hints::hint_implementation::secp::{is_on_curve, read_ec_point_from_address};
 use crate::hints::hint_implementation::state::{
     compute_commitments_on_finalized_state_with_aliases,
-    decode_node,
     guess_classes_ptr,
     guess_state_ptr,
-    load_bottom,
-    load_edge,
     set_preimage_for_class_commitments,
     set_preimage_for_current_commitment_info,
     set_preimage_for_state_commitments,
