+++ crates/starknet_os/src/hints/hint_implementation/state.rs
@@ -5,17 +5,13 @@ use cairo_vm::hint_processor::builtin_hint_processor::hint_utils::{
     insert_value_from_var_name,
 };
 use starknet_api::core::ContractAddress;
-use starknet_patricia::hash::hash_trait::HashOutput;
-use starknet_patricia::patricia_merkle_tree::node_data::inner_node::EdgeData;
 use starknet_types_core::felt::Felt;
 
-use super::patricia::utils::Preimage;
 use crate::hints::error::{OsHintError, OsHintResult};
-use crate::hints::hint_implementation::patricia::utils::{create_preimage_mapping, PreimageMap};
+use crate::hints::hint_implementation::patricia::utils::create_preimage_mapping;
 use crate::hints::types::HintArgs;
 use crate::hints::vars::{CairoStruct, Const, Ids, Scope};
 use crate::io::os_input::CommitmentInfo;
-use crate::vm_utils::{insert_value_to_nested_field, insert_values_to_fields};
 
 #[derive(Copy, Clone)]
 pub(crate) enum CommitmentType {
@@ -138,55 +134,6 @@ pub(crate) fn set_preimage_for_current_commitment_info<S: StateReader>(
         create_preimage_mapping(&commitment_info.commitment_facts)?,
     );
     Ok(())
-}
-
-pub(crate) fn load_edge<S: StateReader>(
-    HintArgs { hint_processor, vm, ids_data, ap_tracking, exec_scopes, .. }: HintArgs<'_, '_, S>,
-) -> OsHintResult {
-    // TODO(Nimrod): Verify that it's ok to ignore the scope variable
-    // `__patricia_skip_validation_runner`.
-    let node = HashOutput(get_integer_from_var_name(Ids::Node.into(), vm, ids_data, ap_tracking)?);
-    let preimage_mapping: &PreimageMap = exec_scopes.get_ref(Scope::Preimage.into())?;
-    let preimage = preimage_mapping.get(&node).ok_or(OsHintError::MissingPreimage(node))?;
-    let Preimage::Edge(EdgeData { bottom_hash, path_to_bottom }) = preimage else {
-        // We expect an edge node.
-        return Err(OsHintError::AssertionFailed {
-            message: format!("An edge node is expected, found {preimage:?}"),
-        });
-    };
-    // Allocate space for the edge node.
-    let edge_ptr = vm.add_memory_segment();
-    insert_value_from_var_name(Ids::Edge.into(), edge_ptr, vm, ids_data, ap_tracking)?;
-    // Fill the node fields.
-    insert_values_to_fields(
-        edge_ptr,
-        CairoStruct::NodeEdge,
-        vm,
-        &[
-            ("length", Felt::from(path_to_bottom.length).into()),
-            ("path", Felt::from(&path_to_bottom.path).into()),
-            ("bottom", bottom_hash.0.into()),
-        ],
-        hint_processor.os_program,
-    )?;
-    let hash_ptr = get_relocatable_from_var_name(Ids::HashPtr.into(), vm, ids_data, ap_tracking)?;
-    insert_value_to_nested_field(
-        hash_ptr,
-        hint_processor.commitment_type.hash_builtin_struct(),
-        vm,
-        &["result"],
-        hint_processor.os_program,
-        node.0 - Felt::from(path_to_bottom.length),
-    )?;
-    Ok(())
-}
-
-pub(crate) fn load_bottom<S: StateReader>(HintArgs { .. }: HintArgs<'_, '_, S>) -> OsHintResult {
-    todo!()
-}
-
-pub(crate) fn decode_node<S: StateReader>(HintArgs { .. }: HintArgs<'_, '_, S>) -> OsHintResult {
-    todo!()
 }
 
 pub(crate) fn guess_state_ptr<S: StateReader>(
